Move cohort processing to a separate file
